+++ src/groups/mwc/mwctst/mwctst_testhelper.cpp
@@ -47,7 +47,7 @@ void TestHelper::printTestName(bslstl::StringRef value)
     }
 
     bsl::cout << "\n" << value << "\n";
-    int length = value.length();
+    long unsigned int length = value.length();
     while (length--) {
         bsl::cout << "=";
     }
